Factorise out read from `_udr` register
This change concerns `HardwareSerial::_rx_complete_irq`, an
interrupt service routine.

This function always reads from `_udr` (a pointer to the `UDR0`
hardware register or equivalent), regardless of whether
`bit_is_clear` returns true or false. Hence, this read can be
performed outside of the `if` block, and the result shared
between both branches (of course, the `else` branch discards
this result anyway).

This change reduces the length of the compiled function by two
instructions (four bytes), measured when compiling with `-Os
-flto`.